Create StudentAlgoNov8.py
Overwrites the code that created the dataframe that creates a specific dataframe for a given program and finds the team numbers they can join.
This PR now takes that concept and expands it to loop through all 10 teams and assigns the avaialble teams to each student (looping through them too).